requirements.txt
dotenv
langchain-google-vertexai
langchain-text-splitters
langchain-community
langchain-nvidia-ai-endpoints
langchain-community
langchain-chroma
faiss-cpu
langgraph
beautifulsoup4
tiktoken
pypdf
ctransformers
sentence_transformers
torch
langchain_huggingface
